Bundle cli with rollup
This PR allows us to eliminate build step before running tests.

lib/cli.js is called directly with `node -r esm`

bin/remarkable just requires compiled version from dist/cjs/cli.js
