+++ package.json
@@ -57,9 +57,10 @@
   },
   "scripts": {
     "build": "make rollup",
-    "test": "make rollup && make test",
-    "test-ci": " make rollup && nyc mocha -r esm -R spec --bail",
-    "coverage": "yarn add coveralls@2 && nyc report --reporter=text-lcov | coveralls"
+    "test": "make test",
+    "test-ci": " nyc mocha -r esm -R spec --bail",
+    "coverage": "yarn add coveralls@2 && nyc report --reporter=text-lcov | coveralls",
+    "prepublishOnly": "make rollup"
   },
   "nyc": {
     "exclude": [
@@ -82,6 +83,7 @@
     "nyc": "^14.1.1",
     "rollup": "^1.16.7",
     "rollup-plugin-commonjs": "^10.0.1",
+    "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-terser": "^5.1.1"
   },
